rename and move function to its own file
This is in preparation to adding a second function.
It also made it possible to simplify the test since the runtime initializeApp can now be skipped by the test.
