+++ functions-v2/test/on-user-doc-written.test.ts
@@ -4,11 +4,7 @@ import {
 } from "firebase-functions-test/lib/providers/firestore";
 import * as logger from "firebase-functions/logger";
 import * as admin from "firebase-admin";
-
-// We cannot import the function here because we need to call
-// initializeFFT first in order to set things up before the
-// initializeApp is called in the function module.
-// import {updateClassDocNetworksOnUserChange} from "../src";
+import {onUserDocWritten} from "../src/on-user-doc-written";
 
 jest.mock("firebase-functions/logger");
 
@@ -16,9 +12,11 @@ process.env["FIRESTORE_EMULATOR_HOST"]="127.0.0.1:8088";
 const projectConfig = {projectId: "demo-test"};
 const fft = initializeFFT(projectConfig);
 
-// We can't initialize the firebase admin here because that
-// can only happen once and the function module needs to do it.
-// admin.initializeApp(projectConfig);
+// When the function is running in the cloud initializeApp is called by index.ts
+// Here we are importing the function's module directly so we can call
+// initializeApp ourselves. This is beneficial since initializeApp needs to
+// be called after initializeFFT above.
+admin.initializeApp();
 
 type CollectionRef = admin.firestore.CollectionReference<
   admin.firestore.DocumentData, admin.firestore.DocumentData
@@ -29,14 +27,14 @@ describe("functions", () => {
     await clearFirestoreData(projectConfig);
   });
 
-  describe("updateClassDocNetworksOnUserChange", () => {
+  describe("onUserDocWritten", () => {
     let classesCollection: CollectionRef;
     let usersCollection: CollectionRef;
 
-    function init() {
+    beforeEach(() => {
       classesCollection = admin.firestore().collection("demo/test/classes");
       usersCollection = admin.firestore().collection("demo/test/users");
-    }
+    });
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     function writeClassDocs(classDocs: any[]) {
@@ -57,17 +55,7 @@ describe("functions", () => {
     }
 
     test("add new network", async () => {
-      // The function module has to be imported after initializeFFT is called.
-      // The initializeFFT sets up environment vars so the
-      // admin.initializeApp() in index.ts will have the same project
-      // settings.
-      const {updateClassDocNetworksOnUserChange} = await import("../src");
-
-      // We can't use beforeEach because this needs to happen after the import.
-      // And we can't put the import in beforeEach because it would be hard to
-      // get the imported function typed properly.
-      init();
-      const wrapped = fft.wrap(updateClassDocNetworksOnUserChange);
+      const wrapped = fft.wrap(onUserDocWritten);
 
       const event = {
         params: {
@@ -144,9 +132,7 @@ describe("functions", () => {
     });
 
     test("remove network", async () => {
-      const {updateClassDocNetworksOnUserChange} = await import("../src");
-      init();
-      const wrapped = fft.wrap(updateClassDocNetworksOnUserChange);
+      const wrapped = fft.wrap(onUserDocWritten);
 
       const event = {
         params: {
@@ -225,9 +211,7 @@ describe("functions", () => {
     // If there is overlap between the networks of the co-teachers then removing
     // a network from one co-teacher might not change the networks of the class
     test("no network change in a class", async () => {
-      const {updateClassDocNetworksOnUserChange} = await import("../src");
-      init();
-      const wrapped = fft.wrap(updateClassDocNetworksOnUserChange);
+      const wrapped = fft.wrap(onUserDocWritten);
 
       const event = {
         params: {
